Don't create summary dataframes if they already exist.

+++ code/04_reformat_output_dataframes.R
@@ -33,33 +33,35 @@ all_output_files <- grep("5trees", all_output_files, value = TRUE, invert = TRUE
 
 #### 3. Prepare summary of the manual topology check csv file ####
 ### Summarise topology results (as percentage of each output topology)
-# Read in .xlsx file with manual topology check results
-topology_check_x_file <- grep("xls", grep("ML_tree_topology_ManualCheck", all_output_files, value = TRUE), value = TRUE)
-topology_check_x_file <- grep("Summary", topology_check_x_file, value = TRUE, invert = TRUE)
-topology_check_df <- as.data.frame(read_excel(path = topology_check_x_file, sheet = "Topology"))
-# Remove Simion and Hejnol datasets - too computationally intensive to run full ML models
-topology_check_df <- topology_check_df[which(topology_check_df$dataset != "Hejnol2009" & topology_check_df$dataset != "Simion2017"), ]
-# Convert "NA" to NA
-topology_check_df$model_BIC[which(topology_check_df$model_BIC == "NA")] <- NA
-topology_check_df$sister_group[which(topology_check_df$sister_group == "NA")] <- NA
-topology_check_df$UFB_support[which(topology_check_df$UFB_support == "NA")] <- NA
-topology_check_df$PORI_topology[which(topology_check_df$PORI_topology == "NA")] <- NA
-topology_check_df$`CTEN+CNID_monophyletic`[which(topology_check_df$`CTEN+CNID_monophyletic` == "NA")] <- NA
-topology_check_df$PLAC_present[which(topology_check_df$PLAC_present == "NA")] <- NA
-topology_check_df$tree_BIC[which(topology_check_df$tree_BIC == "NA")] <- NA
-# Convert necessary columns to numeric
-topology_check_df$model_BIC <- as.numeric(topology_check_df$model_BIC)
-topology_check_df$tree_BIC <- as.numeric(topology_check_df$tree_BIC) 
-# Summarise results for each dataset as a percentage
-dataset_ids <- unique(paste0(topology_check_df$dataset, ".", topology_check_df$matrix_name))
-summary_topology_list <- lapply(dataset_ids, summarise.topology.results, topology_check_df, 
-                                excluded_models = c("C10", "C30", "C40", "C50"))
-summary_topology_df <-  as.data.frame(do.call(rbind, summary_topology_list))
-# Sort output by year
-summary_topology_df <- summary_topology_df[order(summary_topology_df$dataset_year, summary_topology_df$dataset, summary_topology_df$matrix_name),]
-# Write the output
 summary_topology_file <- paste0(output_file_dir, "summary_ML_tree_topology.csv")
-write.csv(summary_topology_df, file = summary_topology_file, row.names = FALSE)
+if (file.exists(summary_topology_file) == FALSE){
+  # Read in .xlsx file with manual topology check results
+  topology_check_x_file <- grep("xls", grep("ML_tree_topology_ManualCheck", all_output_files, value = TRUE), value = TRUE)
+  topology_check_x_file <- grep("Summary", topology_check_x_file, value = TRUE, invert = TRUE)
+  topology_check_df <- as.data.frame(read_excel(path = topology_check_x_file, sheet = "Topology"))
+  # Remove Simion and Hejnol datasets - too computationally intensive to run full ML models
+  topology_check_df <- topology_check_df[which(topology_check_df$dataset != "Hejnol2009" & topology_check_df$dataset != "Simion2017"), ]
+  # Convert "NA" to NA
+  topology_check_df$model_BIC[which(topology_check_df$model_BIC == "NA")] <- NA
+  topology_check_df$sister_group[which(topology_check_df$sister_group == "NA")] <- NA
+  topology_check_df$UFB_support[which(topology_check_df$UFB_support == "NA")] <- NA
+  topology_check_df$PORI_topology[which(topology_check_df$PORI_topology == "NA")] <- NA
+  topology_check_df$`CTEN+CNID_monophyletic`[which(topology_check_df$`CTEN+CNID_monophyletic` == "NA")] <- NA
+  topology_check_df$PLAC_present[which(topology_check_df$PLAC_present == "NA")] <- NA
+  topology_check_df$tree_BIC[which(topology_check_df$tree_BIC == "NA")] <- NA
+  # Convert necessary columns to numeric
+  topology_check_df$model_BIC <- as.numeric(topology_check_df$model_BIC)
+  topology_check_df$tree_BIC <- as.numeric(topology_check_df$tree_BIC) 
+  # Summarise results for each dataset as a percentage
+  dataset_ids <- unique(paste0(topology_check_df$dataset, ".", topology_check_df$matrix_name))
+  summary_topology_list <- lapply(dataset_ids, summarise.topology.results, topology_check_df, 
+                                  excluded_models = c("C10", "C30", "C40", "C50"))
+  summary_topology_df <-  as.data.frame(do.call(rbind, summary_topology_list))
+  # Sort output by year
+  summary_topology_df <- summary_topology_df[order(summary_topology_df$dataset_year, summary_topology_df$dataset, summary_topology_df$matrix_name),]
+  # Write the output
+  write.csv(summary_topology_df, file = summary_topology_file, row.names = FALSE)
+}
 
 ### Nicely format output data frames of tree topologies and of sponge topologies
 model_order <- c("PMSF_C20", "PMSF_C60", "PMSF_LG_C20", "PMSF_LG_C60", 
@@ -67,29 +69,35 @@ model_order <- c("PMSF_C20", "PMSF_C60", "PMSF_LG_C20", "PMSF_LG_C60",
                  "EX2", "EX3", "GTR20", "JTT", "JTTDCMut", "LG", "LG4M",
                  "mtZOA", "PMB", "Poisson", "rtREV", "UL2", "UL3", "WAG",
                  "ModelFinder")
+
+# Order dataframe by dataset/matrix
 ordered_dataset_ids <- paste0(summary_topology_df$dataset, ".", summary_topology_df$matrix_name)
 ## For tree topologies:
-# Extract tree topologies from dataframe
-tree_topology_list <- lapply(ordered_dataset_ids, tree.topology.results, topology_check_df, model_order)
-tree_topology_df <- as.data.frame(do.call(cbind, tree_topology_list))
-# Format dataframe
-names(tree_topology_df) <- ordered_dataset_ids
-tree_topology_df$row_names <- c("dataset", "matrix_name", model_order)
-tree_topology_df <- tree_topology_df[, c("row_names", ordered_dataset_ids)]
-# Output dataframe
 tree_topology_file <- paste0(output_file_dir, "all_models_ML_tree_topology.csv")
-write.csv(tree_topology_df, file = tree_topology_file, row.names = FALSE)
+if (file.exists(tree_topology_file) == FALSE){
+  # Extract tree topologies from dataframe
+  tree_topology_list <- lapply(ordered_dataset_ids, tree.topology.results, topology_check_df, model_order)
+  tree_topology_df <- as.data.frame(do.call(cbind, tree_topology_list))
+  # Format dataframe
+  names(tree_topology_df) <- ordered_dataset_ids
+  tree_topology_df$row_names <- c("dataset", "matrix_name", model_order)
+  tree_topology_df <- tree_topology_df[, c("row_names", ordered_dataset_ids)]
+  # Output dataframe
+  write.csv(tree_topology_df, file = tree_topology_file, row.names = FALSE)
+}
 ## For Porifera topologies:
-# Extract tree topologies from dataframe
-pori_topology_list <- lapply(ordered_dataset_ids, porifera.topology.results, topology_check_df, model_order)
-pori_topology_df <- as.data.frame(do.call(cbind, pori_topology_list))
-# Format dataframe
-names(pori_topology_df) <- ordered_dataset_ids
-pori_topology_df$row_names <- c("dataset", "matrix_name", model_order)
-pori_topology_df <- pori_topology_df[, c("row_names", ordered_dataset_ids)]
-# Output dataframe
 pori_topology_file <- paste0(output_file_dir, "all_models_ML_Porifera_topology.csv")
-write.csv(pori_topology_df, file = pori_topology_file, row.names = FALSE)
+if (file.exists(pori_topology_file) == FALSE){
+  # Extract tree topologies from dataframe
+  pori_topology_list <- lapply(ordered_dataset_ids, porifera.topology.results, topology_check_df, model_order)
+  pori_topology_df <- as.data.frame(do.call(cbind, pori_topology_list))
+  # Format dataframe
+  names(pori_topology_df) <- ordered_dataset_ids
+  pori_topology_df$row_names <- c("dataset", "matrix_name", model_order)
+  pori_topology_df <- pori_topology_df[, c("row_names", ordered_dataset_ids)]
+  # Output dataframe
+  write.csv(pori_topology_df, file = pori_topology_file, row.names = FALSE)
+}
